9/6/24 Reporte Final Ok

+++ ClientePersonaService/ClientePersonaService/src/main/java/com/devsu/ClientePersonaService/model/reporte/CuentaReporte.java
@@ -3,6 +3,7 @@ package com.devsu.ClientePersonaService.model.reporte;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -11,7 +12,7 @@ public class CuentaReporte {
     private Long numeroCuenta;
     private String tipoCuenta;
 
-    private Double saldoInicial;
+    private BigDecimal saldoInicial;
     private Boolean estado;
     @JsonProperty("clienteId")
     private Long clienteId;
