+++ ClientePersonaService/ClientePersonaService/src/main/java/com/devsu/ClientePersonaService/model/reporte/MovimientoReporte.java
@@ -2,6 +2,7 @@ package com.devsu.ClientePersonaService.model.reporte;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -9,7 +10,7 @@ public class MovimientoReporte {
     private Long id;
     private LocalDate fecha;
     private String tipoMovimiento;
-    private Double valor;
-    private Double saldo;
+    private BigDecimal valor;
+    private BigDecimal saldo;
 
 }
