+++ ClientePersonaService/ClientePersonaService/src/main/java/com/devsu/ClientePersonaService/model/reporteFinal/MovimientoReporteFinal.java
@@ -3,6 +3,7 @@ package com.devsu.ClientePersonaService.model.reporteFinal;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -18,8 +19,8 @@ public class MovimientoReporteFinal {
     private String tipoMovimiento;
 
     @JsonProperty("valor")
-    private Double valor;
+    private BigDecimal valor;
 
     @JsonProperty("saldo")
-    private Double saldo;
+    private BigDecimal saldo;
 }
